+++ src/module/MPC/AdaptiveMPC/AdaptiveMPC.cpp
@@ -8,12 +8,12 @@ namespace MPC {
 
         AdaptiveMPC::AdaptiveMPC(Optimizer optimizer) : optimizer(optimizer) {}
 
-        std::pair<bool, bool> AdaptiveMPC::Compute(std::vector<double>& states, double theta) {
+        std::pair<bool, bool> AdaptiveMPC::Compute(std::vector<float>& states, float theta) {
 
             Model::dynamic_model m;
 
             // Convert the input setpoint to the y length setpoint
-            double setpoint = m.radius * theta;
+            float setpoint = m.radius * theta;
 
             return (optimizer.FirstLayer(m, states, setpoint));
         }
